# Better BGZF format streaming performance.