### Checklist
- [x] Pull request details were added to CHANGELOG.rst
- [x] Documentation was updated (if needed)

Integrate CPython's gzip header code which works on streams. 
Adapt for python-isal:
- header corruption check
- Check for BGZF fields in the Extra fields. If found, save the block size and use that to inform the amount that should be read from the next block.

This change was advantageous mainly for bam files:
Before on uncompressed, level 1 and level 5 compressed bams respectively:
```
Benchmark 1: python -m isal.igzip -cd ~/test/nanopore_100000reads.u.bam > /dev/null
  Time (mean ± σ):     613.3 ms ±  25.2 ms    [User: 478.7 ms, System: 134.4 ms]
  Range (min … max):   600.7 ms … 684.3 ms    10 runs
 
Benchmark 1: python -m isal.igzip -cd ~/test/nanopore_100000reads.1.bam > /dev/null
  Time (mean ± σ):      3.003 s ±  0.004 s    [User: 2.901 s, System: 0.101 s]
  Range (min … max):    2.997 s …  3.008 s    10 runs

Benchmark 1: python -m isal.igzip -cd ~/test/nanopore_100000reads.5.bam > /dev/null
  Time (mean ± σ):      3.120 s ±  0.010 s    [User: 3.013 s, System: 0.106 s]
  Range (min … max):    3.105 s …  3.132 s    10 runs
```

After:
```
Benchmark 1: python -m isal.igzip -cd ~/test/nanopore_100000reads.u.bam > /dev/null
  Time (mean ± σ):     448.6 ms ±   8.3 ms    [User: 355.9 ms, System: 92.1 ms]
  Range (min … max):   433.4 ms … 464.0 ms    10 runs

Benchmark 1: python -m isal.igzip -cd ~/test/nanopore_100000reads.1.bam > /dev/null
  Time (mean ± σ):      2.854 s ±  0.014 s    [User: 2.771 s, System: 0.082 s]
  Range (min … max):    2.837 s …  2.876 s    10 runs

Benchmark 1: python -m isal.igzip -cd ~/test/nanopore_100000reads.5.bam > /dev/null
  Time (mean ± σ):      2.943 s ±  0.012 s    [User: 2.866 s, System: 0.076 s]
  Range (min … max):    2.933 s …  2.970 s    10 runs
```

A pretty significant difference proportional to the number of GZIP blocks used. Especially for uncompressed bgzf it is pretty good, which is relevant because samtools view -u outputs uncompressed bgzf rather than raw bam (without bgzf format). So this is pretty useful when piping in streams.

@marcelm FYI, this will be particularly useful for BAM parsing in dnaio. It is also quite an eventful change. I do trust my test suite, but unforeseen bugs may occur. I will keep an eye on the cutadapt issue tracker after releasing this change.